Add epoch vitals sync dag.
