+++ services/airflow/src/dags/datenspende_vitaldata/post_processing/shared.py
@@ -8,6 +8,11 @@ import pandas as pd
 import ramda as R
 from pathos.multiprocessing import Pool
 from psycopg2.sql import SQL, Literal
+from src.lib.dag_helpers import (
+    execute_query_and_return_dataframe,
+    connect_to_db_and_upsert_pandas_dataframe_on_constraint,
+)
+from src.lib.test_helpers import set_env_variable_from_dag_config_if_present
 
 from postgres_helpers import (
     DBContext,
@@ -15,11 +20,6 @@ from postgres_helpers import (
     teardown_db_context,
     query_one_element,
 )
-from src.lib.dag_helpers import (
-    execute_query_and_return_dataframe,
-    connect_to_db_and_upsert_pandas_dataframe_on_constraint,
-)
-from src.lib.test_helpers import set_env_variable_from_dag_config_if_present
 
 
 @R.curry
