+++ services/airflow/src/dags/load_epoch_values/__init__.py
@@ -1,0 +1,3 @@
+from .sync_clickhouse import extract_load_epoch_data
+
+__all__ = ["extract_load_epoch_data"]
